+++ src/openhuman/screen_intelligence/mod.rs
@@ -1,5 +1,6 @@
 //! Screen capture, accessibility automation, and vision summaries (macOS-focused).
 
+pub mod cli;
 pub mod ops;
 mod schemas;
 pub mod server;
